refactor: 패키지 구조 개선 및 파일 이동
