+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -1,6 +1,6 @@
 package com.techcourse.controller;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.mapping.asis.Controller;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
