+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/exception/NoHandlerFoundException.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.exception;
 
 public class NoHandlerFoundException extends Exception {
     public NoHandlerFoundException(String httpMethod, String requestURL) {
