+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/mapping/HandlerMappingRegistry.java
@@ -1,5 +1,6 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.mapping;
 
+import com.interface21.webmvc.servlet.mvc.exception.NoHandlerFoundException;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
