+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/mapping/asis/ForwardController.java
@@ -1,8 +1,7 @@
-package com.interface21.webmvc.servlet.mvc.asis;
+package com.interface21.webmvc.servlet.mvc.mapping.asis;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-
 import java.util.Objects;
 
 public class ForwardController implements Controller {
